only install yay for arch

+++ install/01-yay.sh
@@ -1,10 +1,11 @@
 #!/usr/bin/env bash
-sudo pacman -S --needed --noconfirm base-devel
-
-if ! command -v yay &>/dev/null; then
-  git clone https://aur.archlinux.org/yay-bin.git
-  cd yay-bin
-  makepkg -si --noconfirm
-  cd ~
-  rm -rf yay-bin
+if [ ${FLAVOR} = "arch" ]; then
+    install_packages base-devel
+    if ! command -v yay &>/dev/null; then
+      git clone https://aur.archlinux.org/yay-bin.git
+      cd yay-bin
+      makepkg -si --noconfirm
+      cd ~
+      rm -rf yay-bin
+    fi
 fi
